requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Werkzeug==2.3.7
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
mysqlclient==2.2.0